+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.19]
+
+- added size `flex` to `khao-tab`
+- added `maxWidth` attribute to `khao-tab-bar`
+
+## [1.4.18]
+
+- reduced sizes in `khao-tab`
+
+## [1.4.17]
+
+- added further improvemen for scrolling of active tab in to view in `khao-tab-bar`
+
+## [1.4.16]
+
+- improved scrolling of active tab in to view in `khao-tab-bar`
+
+## [1.4.15]
+
+- added helper class `Breakpoint` and used it in `khao-tab`
+
 ## [1.4.14]
 
 - deprecate `khao-tabs-bar` in favor of `khao-tab-bar`